Fix the products cache and its data model
The cascade rules ran child to parent: deleting an image or a tag
deleted the product it belonged to. update(from:) then deleted every
child on each save, so the second save of a product removed it. The
rules now point parent to child, and update(from:) diffs instead of
rebuilding.

Price moves from Float to Decimal. Money is not representable in binary
floating point, and the storage round trip was changing the value.

id was optional with no uniqueness constraint, which allowed duplicate
rows for the same product. It is now required and unique.

Reads had no cache policy, so getProducts() never refetched once a
single row existed. CachePolicy makes the choice explicit and defaults
to an hour.

+++ Modules/Features/Products/Sources/Domain/Product.swift
@@ -7,7 +7,7 @@ public struct Product: Identifiable, Sendable {
     public let title: String
     public let description: String
     public let category: String
-    public let price: Double
+    public let price: Decimal
     public let tags: [String]
     public let brand: String?
     public let meta: Meta
@@ -19,7 +19,7 @@ public struct Product: Identifiable, Sendable {
         title: String,
         description: String,
         category: String,
-        price: Double,
+        price: Decimal,
         tags: [String],
         brand: String?,
         meta: Meta,
